Reconocimiento por voz/voice_audit_dao/requirements.txt
SpeechRecognition==3.10.4
PyAudio==0.2.14
psycopg2-binary==2.9.9
python-dotenv==1.0.0
